requirements.txt
Django==1.10.5
Markdown==2.6.8
Pillow==4.0.0
PyJWT==1.4.2
PyMySQL==0.7.10
appdirs==1.4.0
braintree==3.35.0
defusedxml==0.5.0
django-allauth==0.30.0
django-appconf==1.0.2
django-ckeditor==5.2.1
django-cors-headers==2.0.2
django-crispy-forms==1.6.1
django-debug-toolbar==1.6
django-el-pagination==3.0.1
django-filter==1.0.1
django-imagekit==4.0
django-registration-redux==1.4
djangorestframework==3.5.4
djangorestframework-jwt==1.9.0
mysqlclient==1.3.9
oauthlib==2.0.1
olefile==0.44
packaging==16.8
pilkit==2.0
pyparsing==2.1.10
python3-openid==3.0.10
requests==2.13.0
requests-oauthlib==0.8.0
six==1.10.0
sqlparse==0.2.3
